Got the Save/Save Draft buttons working in the draft editor.
Updated the C++ warnings: reduced the strict-overlow to 4 instead of 5, tested with -Wconversion and fixed many errors.
"Fixed" comparison between floating point numbers (== and !=).
Fixed the modified_content() signal so it saves the modified info in the data table as it should be.
Added a copy in the new revision so we can create new revisions for users editing work.
Fixed many problems with the token filter.
Removed some debug code.
Added a cool .js (cloud-zoom.js) and will have to create a plugin for it.
Got the POST from the editor.js to work.

+++ libQtCassandra/include/QtCassandra/QCassandraValue.h
@@ -180,8 +180,8 @@ inline int16_t int16Value(const QByteArray& array, const int index = 0)
     || static_cast<unsigned int>(index + 2) > static_cast<unsigned int>(array.size())) {
         throw std::runtime_error("buffer too small for this int16Value");
     }
-    return (static_cast<int16_t>(static_cast<unsigned char>(array.at(index + 0))) << 8)
-         | static_cast<int16_t>(static_cast<unsigned char>(array.at(index + 1)));
+    return static_cast<int16_t>((static_cast<unsigned char>(array.at(index + 0)) << 8)
+                               | static_cast<unsigned char>(array.at(index + 1)));
 }
 
 inline void appendUInt16Value(QByteArray& array, const uint16_t value)
@@ -200,8 +200,8 @@ inline uint16_t uint16Value(const QByteArray& array, const int index = 0)
     || static_cast<unsigned int>(index + 2) > static_cast<unsigned int>(array.size())) {
         throw std::runtime_error("buffer too small for this uint16Value");
     }
-    return (static_cast<uint16_t>(static_cast<unsigned char>(array.at(index + 0))) << 8)
-         | static_cast<uint16_t>(static_cast<unsigned char>(array.at(index + 1)));
+    return static_cast<uint16_t>((static_cast<unsigned char>(array.at(index + 0)) << 8)
+                                | static_cast<unsigned char>(array.at(index + 1)));
 }
 
 // Int32
